Metropolis Hastings Algorithm for MCMCs
Toy example that is hard-coded to work for Gaussians. The function to be sampled from is a standard Gaussian by default: mean is 0 and standard deviation is 1. The proposal distribution is also a Gaussian.

Keeping the sampling function's parameters constant and changing the proposal distribution's parameters is an easy way to gain key intuitions about how the Metropolis-Hastings algorithm works. This can be easily generalized to functions in general, but for didactic purposes, this should suffice.